Split seasonal_sweep into its own dispatcher module
scrape_dispatcher.py was 551 lines housing three dispatchers, and the
seasonal one shares no helpers with the other two — it's purely a
discovery pass that paginates /seasons/now and hands off work to
user_scrape via bulk-inserted child jobs.

Extract seasonal_sweep_dispatcher into its own module. The remaining
file is the catalog-mutation pair (user_scrape + update_sweep) that
genuinely share _apply_media_diff / _weighted_score. main.py imports
the dispatcher from its new location; tests update their monkeypatch
paths to point at the new module. Document the WHY in a comment so a
future agent doesn't optimize the three sequential DAO reads with
asyncio.gather (unsafe on a shared AsyncSession).
