+++ phsar/app/main.py
@@ -26,10 +26,10 @@ from app.services.backup_dispatcher import backup_dispatcher
 from app.services.job_worker import job_worker
 from app.services.merge_detection_service import backfill_merge_candidates
 from app.services.scrape_dispatcher import (
-    seasonal_sweep_dispatcher,
     update_sweep_dispatcher,
     user_scrape_dispatcher,
 )
+from app.services.seasonal_sweep_dispatcher import seasonal_sweep_dispatcher
 
 setup_logging()
 logger = logging.getLogger(__name__)
